+++ test/libsolidity/syntaxTests/array/uninitialized_calldata_var.sol
@@ -1,0 +1,9 @@
+contract C {
+	function f() public {
+		uint[] calldata x;
+		uint[10] calldata y;
+	}
+}
+// ----
+// DeclarationError: (38-55): Uninitialized calldata pointer.
+// DeclarationError: (59-78): Uninitialized calldata pointer.